+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 0.9.dev.0
 ---------
 
+- Add "Cash and Carry" shipping.
+  [rnix]
+
 - To enable and disable features on contexts (like vendor areas), don't use
   ``directlyProvides`` and erase all other directly provided interfaces, but
   use ``alsoProvides``. That's needed for collective.lineage sites to work